+++ pkg/prom/instance/noop.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/prometheus/prometheus/scrape"
+	"github.com/prometheus/prometheus/storage"
 )
 
 // NoOpInstance implements the Instance interface in pkg/prom
@@ -29,4 +30,9 @@ func (NoOpInstance) TargetsActive() map[string][]*scrape.Target {
 // StorageDirectory implements Instance.
 func (NoOpInstance) StorageDirectory() string {
 	return ""
+}
+
+// Appender implements Instance
+func (NoOpInstance) Appender(_ context.Context) storage.Appender {
+	return nil
 }